refactor: move database ids to events
