+++ common/src/events/types/index.ts
@@ -1,3 +1,5 @@
+export * from "./match-database-ids.schema";
+export * from "./scrim-database-ids.schema";
 export * from "./skill-group-changed.schema";
 export * from "./submission.schemas";
 export * from "./team-changed.schema";
